+++ src/main/java/com/ohdaesan/shallwepets/auth/filter/JwtAuthorizationFilter.java
@@ -71,6 +71,7 @@ public class JwtAuthorizationFilter extends BasicAuthenticationFilter {
                 "/bookmark/(.*)/(.*)",
                 "/bookmark/delete",
                 "/points/(.*)",
+                "/map",
                 "/swagger-ui/(.*)",          // swagger 설정
                 "/swagger-ui/index.html",    // swagger 설정
                 "/v3/api-docs",              // swagger 설정
