--- src/main/java/com/ohdaesan/shallwepets/map/model/Business.java
@@ -1,24 +1,0 @@
-package com.ohdaesan.shallwepets.map.model;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import org.springframework.data.relational.core.mapping.Table;
-
-@Table(name = "business")
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-@Entity
-public class Business {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long postNo;
-    private String fcltyNm;
-    private String rdnmadrNm;
-    private String telNo;
-}
