--- src/main/java/com/ohdaesan/shallwepets/map/repository/BusinessRepository.java
@@ -1,7 +1,0 @@
-package com.ohdaesan.shallwepets.map.repository;
-
-import com.ohdaesan.shallwepets.map.model.Business;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface BusinessRepository extends JpaRepository<Business, Long> {
-}
